SplitErrorCorrectionInConnectomics/requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
appdirs==1.4.4
async-timeout==4.0.2
attrs==22.2.0
Bottleneck @ file:///opt/conda/conda-bld/bottleneck_1657175564434/work
brotlipy==0.7.0
certifi @ file:///croot/certifi_1671487769961/work/certifi
cffi @ file:///tmp/abs_98z5h56wf8/croots/recipe/cffi_1659598650955/work
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
click==8.1.3
contourpy==1.0.7
cryptography @ file:///croot/cryptography_1673298753778/work
cycler==0.11.0
docker-pycreds==0.4.0
filelock==3.9.0
flit_core @ file:///opt/conda/conda-bld/flit-core_1644941570762/work/source/flit_core
fonttools==4.38.0
frozenlist==1.3.3
fsspec==2023.1.0
gitdb==4.0.10
GitPython==3.1.30
h5py==3.8.0
huggingface-hub==0.12.0
idna @ file:///croot/idna_1666125576474/work
joblib==1.2.0
kiwisolver==1.4.4
lightning-utilities==0.6.0.post0
matplotlib==3.6.3
mkl-fft==1.3.1
mkl-random @ file:///tmp/build/80754af9/mkl_random_1626186064646/work
mkl-service==2.4.0
multidict==6.0.4
numexpr @ file:///croot/numexpr_1668713893690/work
numpy @ file:///croot/numpy_and_numpy_base_1672336185480/work
packaging==23.0
pandas==1.5.2
pathtools==0.1.2
Pillow==9.3.0
protobuf==4.21.12
psutil==5.9.4
ptflops==0.6.9
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pyignite==0.6.0
pyOpenSSL @ file:///opt/conda/conda-bld/pyopenssl_1643788558760/work
pyparsing==3.0.9
PySocks @ file:///tmp/build/80754af9/pysocks_1605305779399/work
python-dateutil @ file:///tmp/build/80754af9/python-dateutil_1626374649649/work
pytorch-ignite==0.4.10
pytorch-lightning==1.9.0
pytz @ file:///croot/pytz_1671697431263/work
PyYAML==6.0
requests @ file:///opt/conda/conda-bld/requests_1657734628632/work
scikit-learn==1.2.1
scikit-plot==0.3.7
scipy==1.10.0
sentry-sdk==1.14.0
setproctitle==1.3.2
six @ file:///tmp/build/80754af9/six_1644875935023/work
smmap==5.0.0
threadpoolctl==3.1.0
timm==0.6.12
torch==1.13.1
torchaudio==0.13.1
torchmetrics==0.11.1
torchvision==0.14.1
tqdm==4.64.1
typing_extensions @ file:///croot/typing_extensions_1669924550328/work
urllib3 @ file:///croot/urllib3_1673575502006/work
wandb==0.13.9
yarl==1.8.2
